pyproject.toml
[tool.poetry]
name = "datateam"
version = "0.1.0"
description = ""
authors = ["cristinabaldor <ccbaldor@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
jupyter-book = "^0.12.0"
ghp-import = "^2.0.2"
pivottablejs = "^0.9.0"
pandasgui = "^0.2.13"
qgrid = "^1.3.1"
jupyterlab = "^3.2.1"
ipywidgets = "^7.6.5"
tabloo = "^0.0.15"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
